requirements.txt
botocore==1.24.13
jmespath==0.10.0
pyactiveresource==2.2.2
PyJWT==2.3.0
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.5.2
ShopifyAPI==10.0.0
six==1.16.0
urllib3==1.26.8
